+++ includes/rest/class-actors-inbox-controller.php
@@ -176,9 +176,6 @@ class Actors_Inbox_Controller extends Actors_Controller {
 			 */
 			do_action( 'activitypub_rest_inbox_disallowed', $data, $user_id, $type, $activity );
 		} else {
-			// FEP-8fcf: Process Collection-Synchronization header if present.
-			$this->process_collection_synchronization( $request, $data, $user_id );
-
 			/**
 			 * ActivityPub inbox action.
 			 *
